+++ src/components/Group/GroupSelectInteractive.js
@@ -1,5 +1,5 @@
 import { useSelector } from "react-redux"
-import { TextInput } from "./TextInput"
+import { TextInput } from "../TextInput"
 import { useState } from "react"
 import ReactSelect from 'react-select';
 import { GroupTasks } from "./GroupTasks";
